some refactoring. energies folder
