+++ simkit/energies/arap_energy.py
@@ -1,8 +1,8 @@
 import numpy as np
 
-from .volume import volume
-from .deformation_jacobian import deformation_jacobian
-from .polar_svd import polar_svd
+from simkit import volume
+from simkit import deformation_jacobian
+from simkit import polar_svd
 
 # def arap_energy(X, T, U=None, mu=1, vol=None):
 #     if U is None:
