+++ simkit/energies/elastic_energy.py
