+++ simkit/energies/elastic_gradient.py
