+++ simkit/energies/elastic_hessian.py
