+++ simkit/energies/kinetic_energy.py
