+++ simkit/energies/kinetic_gradient.py
