+++ simkit/energies/kinetic_hessian.py
