+++ simkit/energies/quadratic_energy.py
