+++ simkit/energies/quadratic_gradient.py
