+++ simkit/energies/quadratic_hessian.py
