+++ simkit/filesystem/__init__.py
@@ -1,0 +1,1 @@
+from .get_data_directory import get_data_directory
