+++ simkit/filesystem/get_data_directory.py
@@ -1,0 +1,4 @@
+import os
+
+def get_data_directory():
+    return os.path.join(os.path.dirname(__file__), "../../", "data/")
